Move image list in sky browser to its own component
